REFACTOR reactify
Reactjs ftw! Why didn't I use it from the start, to begin with?

This basically does the exact same thing than previously, but it's now
way easier to read, more performant and it fixes small glitches (like
what is displayed when there are actually collections, but none is shown
due to "filter save me later collection" filter).

How, and discovered coffee-react [1] for the occasion, it rocks.

[1] https://github.com/jsdf/coffee-react