REFACTOR focus back PH tab avec closing oauth tab
After oauth is authorized, background page will close hunting-cabin
page, and just leave user on a random tab (the last one in tab list).

Well, *I* know that means you're ready to use the extension. Other
people probably won't.

Since I was forced to use the "tabs" permission to have the current tab
url and see if it's a product page, let's take advantage of that:
background page will keep track of the product hunt tab where extension
was click, open the hunting-cabin page for oauth, close it when done and
focus back the PH tab.